fixed error in gogyo package about package loadgogyo pakcage was load gogyo package and use this.. ummm..;;
