+++ main.go
@@ -2,12 +2,34 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"strings"
 	"time"
 
 	"github.com/DSMdongly/gogyo"
-	"github.com/DSMdongly/gosirak/gosirak"
 )
+
+func ShowDailyMenu(school gogyo.School, year int, month int, day int) {
+	fmt.Printf("\n[ %2d/%2d/%2d ]\n", year, month, day)
+
+	daily := school.GetDailyMenu(year, month, day)
+	times := [3]string{"조식", "중식", "석식"}
+
+	for index := 0; index < 3; index++ {
+		meals := daily[index]
+
+		if meals == nil {
+			continue
+		}
+
+		time := times[index]
+		fmt.Printf("\n[ %s ]\n\n", time)
+
+		for _, meal := range meals {
+			fmt.Printf("- %s\n", meal.Name)
+		}
+	}
+}
 
 func main() {
 	now := time.Now()
@@ -22,10 +44,10 @@ func main() {
 
 	flag.Parse()
 
-	kind := gosirak.GetSchoolKind(*schoolKind)
-	region := gosirak.GetSchoolRegion(*schoolRegion)
+	kind := gogyo.GetSchoolKind(*schoolKind)
+	region := gogyo.GetSchoolRegion(*schoolRegion)
 	code := gogyo.SchoolCode(strings.ToUpper(*schoolCode))
 
 	school := gogyo.NewSchool(kind, region, code)
-	gosirak.ShowDailyMenu(*school, *year, *month, *day)
+	ShowDailyMenu(*school, *year, *month, *day)
 }
